CNAM-257-patient-filters

+++ src/main/scala/fr/polytechnique/cmap/cnam/etl/extractors/patients/IrBenPatients.scala
@@ -29,7 +29,7 @@ private[patients] object IrBenPatients {
       val result = data
         .select(
           col("patientID"),
-          col("BEN_SEX_COD").as("gender")
+          col("BEN_SEX_COD").cast("int").as("gender")
         ).distinct
 
       val patients = result.select(col("patientID")).distinct()
